requirements.txt
beautifulsoup4==4.12.2
certifi==2022.9.14
fpdf==1.7.2
fpdf2==2.7.5
joblib==1.1.0
pybind11==2.11.1
Cython==3.0.2
basemap==1.3.8 
kneed==0.8.3
matplotlib==3.8.0
netCDF4==1.6.3
numpy==1.26.0
pyproj==3.4.1
pymannkendall==1.4.3
PyYAML==6.0
requests==2.28.1
scipy==1.11.2
setuptools==63.4.1
urllib3==1.26.11
